add docs, minor refactorings
